requirements.txt
# Sermon Scribe Dependencies

# Video download
yt-dlp

# Transcription
openai-whisper

# AI processing (segmentation & cleanup)
openai

# YouTube monitoring
google-api-python-client

# Utilities
pyyaml
python-dotenv
